Refactor dns registration
