+++ test/withPipeline/BaseCnpPipelineTest.groovy
@@ -48,6 +48,7 @@ abstract class BaseCnpPipelineTest extends BasePipelineTest {
     helper.registerAllowedMethod("lock", [String.class, Closure.class], null)
     helper.registerAllowedMethod("scmServiceRegistration", [String.class], {})
     helper.registerAllowedMethod("scmServiceRegistration", [String.class, String.class], {})
+    helper.registerAllowedMethod("registerDns", [LinkedHashMap.class], {})
     helper.registerAllowedMethod("wrap", [LinkedHashMap, Closure.class], null)
     helper.registerAllowedMethod("sh", [Map.class], { m ->
       if (m.get('script') == 'pwd') {
