+++ vars/helmPublish.groovy
@@ -1,8 +1,6 @@
 import uk.gov.hmcts.contino.Helm
 
 def call(Map params) {
-  AppPipelineConfig config = params.appPipelineConfig
-
   withAksClient(params.subscriptionName, params.environmentName) {
 
     String chartName = "${params.product}-${params.component}"
@@ -33,13 +31,7 @@ def call(Map params) {
       }
 
       helm.publishIfNotExists(values)
-
-      if (!config.legacyDeployment) {
-        consul.registerDns("${params.product}-${params.component}-${params.environmentName}", ingressIP)
-      }
-      if (config.aksStagingDeployment) {
-        consul.registerDns("${params.product}-${params.component}", ingressIP)
-      }
     }
   }
+  registerDns(params)
 }
