+++ vars/sectionDeployToAKS.groovy
@@ -87,6 +87,8 @@ def call(params) {
     }
 
     if ((config.deployToAKS || config.installCharts) && config.serviceApp) {
+      registerDns(params)
+      
       withSubscription(subscription) {
         withTeamSecrets(config, environment) {
           stage("Smoke Test - AKS ${environment}") {
